+++ src/protocol_level_tokens/token_client.rs
@@ -351,7 +351,7 @@ impl TokenClient {
     }
 
     /// Removes an accounts from the allow list of a token.
-    ///     
+    ///
     /// # Arguments
     ///
     /// * `signer` - A [`WalletAccount`] who's address is used as a sender and
@@ -789,7 +789,7 @@ impl TokenClient {
             expiry,
             token_id,
             operations,
-        )?;
+        );
         let block_item = BlockItem::AccountTransaction(transaction);
         Ok(self.client.send_block_item(&block_item).await?)
     }
@@ -965,11 +965,7 @@ mod tests {
                     deny_list,
                     additional: HashMap::new(),
                 })
-                .map(|s| {
-                    cbor::cbor_encode(&s)
-                        .expect("TokenModuleAccountState must be serializbale")
-                        .into()
-                });
+                .map(|s| cbor::cbor_encode(&s).into());
 
             MockInfo {
                 info: vec![AccountToken {