+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+- Moved the `Upward`,`TokenEvent`, `TokenEventDetails` from the `concordium-base`.
+- Moved the `TokenModuleEvent` from the `concordium-base` as `EncodedTokenModuleEvent`.
+- Moved the `TokenModuleRejectReason` from the `concordium-base` as `EncodedTokenModuleRejectReason`.
+
 ## 9.0.0
 
 - Support for `ProtocolVersion::P10`.