+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Moved the `TokenModuleEvent` from the `concordium-base` as `EncodedTokenModuleEvent`.
 - Moved the `TokenModuleRejectReason` from the `concordium-base` as `EncodedTokenModuleRejectReason`.
 
+## 9.0.1
+
+- Fixes a bug where using `serde` to convert `BlockItemSummary` would result in data loss with regards to transaction sponsor details.
+
 ## 9.0.0
 
 - Support for `ProtocolVersion::P10`.